added method to get tickets for a project and integrated this into the method that gets all projects for a user

+++ app/src/main/java/com/tracker/app/ticket/TicketController.java
@@ -24,7 +24,7 @@ public class TicketController {
     }
 
     @GetMapping("/getproject/{id}")
-    public List<Ticket> getTicketsForProject(@PathVariable int id){
+    public List<Ticket> getTicketsForProject(@PathVariable Long id){
         return ticketService.getTicketsForProject(id);
     }
